requirements.txt
fastapi
uvicorn
aiofiles

fastapi_chameleon
starlette==0.14.2
python-multipart
pymongo[srv]
python-dotenv
passlib
gunicorn
sqlmodel
pytz
psycopg2-binary